Intel ES-RAPP code changes
Code changes related to Intel Energy Saving rAPP.

This rAPP is designed for a co-located cell topology only. Where only the capacity cell can be switched on/off and the macro cell remains turned on.

Changes contain cell on/off decision making logic of rAPP. The rAPP maintains a small history of observed prb usage and queries the load predictor for the future prediction. Based on this prediction a decision to switch on or switch off a cell is taken.